api/requirements.txt
flask==3.0.3
requests==2.31.0   # pour appeler le service recognition
psycopg2-binary==2.9.9  # connexion PostgreSQL
   